adt-prototype/requirements.txt
google-adk
PyGithub
python-dotenv
requests
toml